COM-3840: use components
