+++ src/components/AttendanceSheetDataSelectionForm/styles.ts
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+import { MARGIN } from '../../styles/metrics';
+import { FIRA_SANS_BOLD } from '../../styles/fonts';
+
+const styles = StyleSheet.create({
+  container: {
+    flexGrow: 1,
+    marginHorizontal: MARGIN.LG,
+  },
+  title: {
+    ...FIRA_SANS_BOLD.LG,
+    marginVertical: MARGIN.LG,
+  },
+  button: {
+    marginHorizontal: MARGIN.MD,
+  },
+});
+
+export default styles;
