+++ src/components/UploadMethods/styles.ts
@@ -1,0 +1,10 @@
+import { StyleSheet } from 'react-native';
+import { MARGIN } from '../../styles/metrics';
+
+const styles = StyleSheet.create({
+  button: {
+    marginHorizontal: MARGIN.MD,
+  },
+});
+
+export default styles;
